我的问题

+++ src/main/java/com/hhy/bbs/mapper/UserMapper.java
@@ -14,5 +14,5 @@ public interface UserMapper {
     @Select("select * from user where token=#{token}")
     User findByToken(@Param("token") String token);
     @Select("select * from user where id=#{id}")
-    User findById(@Param("id")Long id);
+    User findById(@Param("id")Integer id);
 }
